splitted OptionsDefinitions by extracting the options print part to separate class
