+++ common/src/main/java/net/adoptopenjdk/icedteaweb/option/OptionsDefinitions.java
@@ -34,21 +34,13 @@ this exception to your version of the library, but you are not
 obligated to do so.  If you do not wish to do so, delete this
 exception statement from your version.
 */
-package net.sourceforge.jnlp;
-
-import net.sourceforge.jnlp.util.docprovider.TextsProvider;
-
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
+package net.adoptopenjdk.icedteaweb.option;
 
 import static net.sourceforge.jnlp.runtime.Translator.R;
 
 public class OptionsDefinitions {
 
-    public static enum OPTIONS {
-
+    public enum OPTIONS {
         //javaws undocummented swithces
         TRUSTALL("-Xtrustall","BOTrustall"),
         //javaws control-options
@@ -163,102 +155,4 @@ public class OptionsDefinitions {
             return R(messageKey);
         }
     }
-
-    public static List<OPTIONS> getItwsettingsCommands() {
-        return Arrays.asList(new OPTIONS[]{
-            OPTIONS.HELP2,
-            OPTIONS.LIST,
-            OPTIONS.GET,
-            OPTIONS.INFO,
-            OPTIONS.SET,
-            OPTIONS.RESET,
-            OPTIONS.RESETALL,
-            OPTIONS.HEADLESS,
-            OPTIONS.CHECK,
-            OPTIONS.VERBOSE
-        });
-    }
-
-    public static List<OPTIONS> getPolicyEditorOptions() {
-        return Arrays.asList(new OPTIONS[]{
-            OPTIONS.HELP1,
-            OPTIONS.FILE,
-            OPTIONS.DEFAULTFILE,
-            OPTIONS.CODEBASE,
-            OPTIONS.SIGNEDBY,
-            OPTIONS.PRINCIPALS,
-            OPTIONS.VERBOSE
-            }
-        );
-    }
-
-    public static List<OPTIONS> getJavaWsControlOptions() {
-        return Arrays.asList(new OPTIONS[]{
-            OPTIONS.ABOUT,
-            OPTIONS.VIEWER,
-            OPTIONS.CLEARCACHE,
-            OPTIONS.LISTCACHEIDS,
-            OPTIONS.LICENSE,
-            OPTIONS.HELP1}
-        );
-    }
-
-    public static List<OPTIONS> getJavaWsRuntimeOptions() {
-        return Arrays.asList(new OPTIONS[]{
-            OPTIONS.VERSION,
-            OPTIONS.ARG,
-            OPTIONS.PARAM,
-            OPTIONS.PROPERTY,
-            OPTIONS.UPDATE,
-            OPTIONS.VERBOSE,
-            OPTIONS.NOSEC,
-            OPTIONS.NOUPDATE,
-            OPTIONS.HEADLESS,
-            OPTIONS.STRICT,
-            OPTIONS.XML,
-            OPTIONS.REDIRECT,
-            OPTIONS.NOFORK,
-            OPTIONS.NOHEADERS,
-            OPTIONS.OFFLINE,
-            OPTIONS.TRUSTNONE,
-            OPTIONS.JNLP,
-            OPTIONS.HTML,
-            OPTIONS.BROWSER
-        });
-    }
-
-    public static List<OPTIONS> getJavaWsOptions() {
-        List<OPTIONS> l = new ArrayList<>();
-        l.addAll(getJavaWsRuntimeOptions());
-        l.addAll(getJavaWsControlOptions());
-        //trustall is not returned by getJavaWsRuntimeOptions
-        //or getJavaWsControlOptions, as it is not desired in documentation
-        l.add(OPTIONS.TRUSTALL);
-        return l;
-    }
-
-    public static void main(String[] args) throws IOException {
-        switch (args[0]) {
-            case TextsProvider.JAVAWS:
-                printOptions(getJavaWsOptions());
-                break;
-            case TextsProvider.ITWEB_SETTINGS:
-                printOptions(getItwsettingsCommands());
-                break;
-            case TextsProvider.POLICY_EDITOR:
-                printOptions(getPolicyEditorOptions());
-                break;
-            default:
-                break;
-        }
-    }
-
-    private static void printOptions(List<OPTIONS> options) {
-        StringBuilder sb = new StringBuilder();
-        for (OPTIONS option : options) {
-            sb.append(option.option).append(" ");
-        }
-        System.out.println(sb.toString().trim());
-    }
-
 }
