+++ core/src/main/java/net/sourceforge/jnlp/cache/ResourceDownloader.java
@@ -1,7 +1,7 @@
 package net.sourceforge.jnlp.cache;
 
+import net.adoptopenjdk.icedteaweb.option.OptionsDefinitions;
 import net.sourceforge.jnlp.DownloadOptions;
-import net.sourceforge.jnlp.OptionsDefinitions;
 import net.sourceforge.jnlp.Version;
 import net.sourceforge.jnlp.runtime.Boot;
 import net.sourceforge.jnlp.runtime.JNLPRuntime;
