+++ core/src/main/java/net/sourceforge/jnlp/util/docprovider/JavaWsTextsProvider.java
@@ -37,7 +37,8 @@ exception statement from your version.
 
 package net.sourceforge.jnlp.util.docprovider;
 
-import net.sourceforge.jnlp.OptionsDefinitions;
+import net.adoptopenjdk.icedteaweb.option.OptionsDefinitions;
+import net.sourceforge.jnlp.util.OptionsDefinitionsPrinter;
 import net.sourceforge.jnlp.config.PathsAndFiles;
 import net.sourceforge.jnlp.runtime.Translator;
 import net.sourceforge.jnlp.util.docprovider.formatters.formatters.Formatter;
@@ -87,9 +88,9 @@ public class JavaWsTextsProvider extends TextsProvider {
         String add3 = Translator.R("JWSoptionsL3");
         String adds = getFormatter().wrapParagraph(add1 + getFormatter().getNewLine() + add2 + getFormatter().getNewLine() + add3);
         String runtime = getFormatter().getBold(Translator.R("JWSoptionsTitle1")) + getFormatter().getNewLine()
-                + optionsToString(OptionsDefinitions.getJavaWsRuntimeOptions());
+                + optionsToString(OptionsDefinitionsPrinter.getJavaWsRuntimeOptions());
         String control = getFormatter().getBold(Translator.R("JWSoptionsTitle2")) + getFormatter().getNewLine()
-                + optionsToString(OptionsDefinitions.getJavaWsControlOptions());
+                + optionsToString(OptionsDefinitionsPrinter.getJavaWsControlOptions());
         return title + adds + getFormatter().wrapParagraph(control) + getFormatter().wrapParagraph(runtime);
     }
